requirements.txt
Django
djangorestframework
django-filter
django-extra-fields
django-cors-headers
isort
flake8
json_log_formatter